refactor: remove api, move components into its own sub folder
