+++ src/pages/app/students.tsx
@@ -1,4 +1,4 @@
-import { DashboardLayout } from '@/containers/DashboardLayout'
+import { DashboardLayout } from '@/containers/dashboard/DashboardLayout'
 
 export default function StudentsPage() {
   return (
